feat(scoring): typed feature engine replaces heuristic score homes
Move path/text classification and destination facts into Html2rss::Scoring
with an explicit feature registry, and delete LinkHeuristics.
