+++ lib/html2rss/scoring.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Html2rss
+  ##
+  # Declarative feature scoring for AutoSource SST segments.
+  module Scoring
+  end
+end
