+++ lib/html2rss/scoring/link_resolver.rb
@@ -1,0 +1,71 @@
+# frozen_string_literal: true
+
+module Html2rss
+  module Scoring
+    ##
+    # Resolves href strings into memoized DestinationFacts for one page base URL.
+    class LinkResolver
+      HREF_BASE_PATTERN = /\A([^#]*)/
+
+      # @param base_url [String, Html2rss::Url]
+      def initialize(base_url)
+        @base_url = base_url
+        @by_href = {}
+        @by_node = {}.compare_by_identity
+        @text_classifier = TextClassifier.new
+      end
+
+      # @return [TextClassifier]
+      attr_reader :text_classifier
+
+      ##
+      # @param node_or_href [SST::Node, String, #to_s]
+      # @return [DestinationFacts, nil]
+      def destination_facts(node_or_href)
+        return @by_node[node_or_href] if node_or_href.is_a?(SST::Node) && @by_node.key?(node_or_href)
+
+        href = extract_href(node_or_href)
+        return unless href
+
+        facts = (@by_href[href] ||= build_facts(href))
+        @by_node[node_or_href] = facts if node_or_href.is_a?(SST::Node)
+        facts
+      rescue ArgumentError
+        nil
+      end
+
+      ##
+      # @param text [String, #to_s]
+      # @return [Boolean]
+      def utility_text?(text) = @text_classifier.utility?(text)
+
+      ##
+      # @param text [String, #to_s]
+      # @return [Boolean]
+      def utility_prefix_text?(text) = @text_classifier.utility_prefix?(text)
+
+      ##
+      # @param text [String, #to_s]
+      # @return [Boolean]
+      def recommended_text?(text) = @text_classifier.recommended?(text)
+
+      private
+
+      def extract_href(node_or_href)
+        raw = case node_or_href
+              when SST::Node then node_or_href.attrs.href
+              else node_or_href
+              end
+        return unless raw
+
+        base = raw.to_s[HREF_BASE_PATTERN, 1].to_s.strip
+        base unless base.empty?
+      end
+
+      def build_facts(href)
+        url = Html2rss::Url.from_relative(href, @base_url)
+        DestinationFacts.build(url)
+      end
+    end
+  end
+end
